Split streamlist specific component from common_functions
